Test_for_team_AI_step1/requirements.txt
google-genai
scenedetect
opencv-python-headless
torch
torchvision
faster-whisper
sentence-transformers
transformers
numpy
pandas
psycopg2-binary
SQLAlchemy>=2.0
pgvector
python-dotenv
tqdm
typer
